requirement.txt
threaded
mysql-connector-python